Ensure correct binary tool dependencies
Running go get in the istio-operator module uses the existing go.mod
and go.sum to build the binary tool. So, the resulting binary depends
on which istio-operator commit it was built on, even though the
binary tool version is properly tracked.

This commit fixes the issue by always running the `go get` command
in a pristine go module, so the resulting binary should only depend
on the contents of the binary tool repo.