Increase timeout of waiting for meshgateway
The order of the two tests seem to matter: the meshgateway and httpbin
pods are started much slower right after the recreation of the istio
resource.

If the meshgateway test is the first one to be executed, than 30s is
enough, but as a second test, it needs 75+ seconds until the
meshgateway address is available.

It seems that nothing is progressing for about a minute right after an
Istio resource is deleted and recreated.